编辑考题 判断+选择防止空指针

+++ src/main/java/com/netEdu/questionPool/question/service/impl/QuestionImpl.java
@@ -36,19 +36,25 @@ public class QuestionImpl implements QuestionService{
     public void update(Question question){
         if (question.getQuestion_id() == 0){
             questionMapper.insertSelective(question);
-            String[] options = question.getOptions().split(",");
-            for (int i = 0;i < options.length;i ++){
-                questionMapper.insertIntoOptions(question.getQuestion_id(),options[i].toString());
+            //如果是选择题
+            if (question.getQuestion_type().equals("1")){
+                String[] options = question.getOptions().split(",");
+                for (int i = 0;i < options.length;i ++){
+                    questionMapper.insertIntoOptions(question.getQuestion_id(),options[i].toString());
+                }
             }
 //            int option_id = questionMapper.selectAnswerForQuestion(question.getQuestion_id(),question.getQuestion_answer());
 //            questionMapper.selectIdToAnswer(question.getQuestion_id(),option_id);
         }
         else {
             questionMapper.updateByPrimaryKeySelective(question);
-            questionMapper.deletePerviousOption(question.getQuestion_id());
-            String[] options = question.getOptions().split(",");
-            for (int i = 0;i < options.length;i ++){
-                questionMapper.insertIntoOptions(question.getQuestion_id(),options[i].toString());
+            //如果是选择题
+            if (question.getQuestion_type().equals("1")){
+                questionMapper.deletePerviousOption(question.getQuestion_id());
+                String[] options = question.getOptions().split(",");
+                for (int i = 0;i < options.length;i ++){
+                    questionMapper.insertIntoOptions(question.getQuestion_id(),options[i].toString());
+                }
             }
 //            int option_id = questionMapper.selectAnswerForQuestion(question.getQuestion_id(),question.getQuestion_answer());
 //            questionMapper.selectIdToAnswer(question.getQuestion_id(),option_id);
